remove PATH attack vulnerability, don't use system, refactor
Executing the previous version of nonet was vulnerable to a path attack when using the --lo option.

  PATH="/evil/:$PATH" nonet --lo echo hello

This would have executed 'ip' as root from /evil/ rather than /sbin/

I changed it to use an absolute path, and avoided 'system' since it shells out to execute the given command.

This meant manually fork+exec'ing and all the additional complexity that brings.